Added template class for multidimenional array

+++ include/Array1D.h
@@ -1,24 +1,27 @@
 #ifndef TABLE1D__H
 #define TABLE1D__H value
+
+#include <cstdlib>
+
 class Array1D
 {
 public:
 	Array1D();
 	~Array1D();
-	void Init(unsigned SizeX);
+	void Init(size_t SizeX);
 	void Clear();
 	inline bool Ready() const{return (data!=0);}
 	double Eval(double ix) const;
 	
-	inline unsigned Len()   const {return sizeX;}
-	inline unsigned SizeX() const {return sizeX;}
+	inline size_t Len()   const {return sizeX;}
+	inline size_t SizeX() const {return sizeX;}
 
-	inline double& operator ()(unsigned i)      {return At(i);}
-	inline double  operator ()(unsigned i) const{return At(i);}
+	inline double& operator ()(size_t i)      {return At(i);}
+	inline double  operator ()(size_t i) const{return At(i);}
 
 private:
-	double& At(unsigned i) const;
-	unsigned sizeX;
+	double& At(size_t i) const;
+	size_t sizeX;
 	double * data;
 };
 #endif
